requirements.txt
nats-py==2.9.0
python-dateutil==2.9.0.post0
